# rms-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RMS API Documentation",
        "description": "API du Revenu Minimum Solidaire (RMS) : auth, portefeuille, file FIFO, rotations, timeline.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscrire un nouveau membre RMS",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "first_name",
                                    "last_name"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+221770000001"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Fatou"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Diop"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "SN"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Dakar"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec état et wallet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+221770000001"
                                                        },
                                                        "first_name": {
                                                            "type": "string",
                                                            "example": "Fatou"
                                                        },
                                                        "last_name": {
                                                            "type": "string",
                                                            "example": "Diop"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/auth/request-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Demander un code OTP par SMS",
                "operationId": "10dbaf9fdc7345c44ec7c44e73d10b35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+221770000001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP généré et (simulé) envoyé"
                    },
                    "429": {
                        "description": "Trop de demandes d'OTP pour ce numéro"
                    }
                }
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Vérifier un code OTP et se connecter",
                "operationId": "abcb5741589707b15c7ecac7a67c13aa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+221770000001"
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP valide, utilisateur connecté"
                    },
                    "400": {
                        "description": "OTP invalide ou expiré"
                    }
                }
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Lister les notifications de l'utilisateur",
                "operationId": "8ad582a11169ade716521701fb684c82",
                "responses": {
                    "200": {
                        "description": "Liste des notifications"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/notifications/mark-as-read": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Marquer des notifications comme lues",
                "operationId": "969842f2d83bbc00ced7966269572d8d",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "uuid-1",
                                            "uuid-2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notifications mises à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}